feat: Refactor ShippingHistoryModal component structure
- Moved ShippingHistoryModal.vue to a new directory structure.
- Created separate components for Filters, Footer, and HistoryTable to enhance modularity and maintainability.
- Updated imports and adjusted the logic to accommodate the new component structure.
- Removed the old ShippingHistoryModal.vue file and its associated logic.
